+++ src/main/kotlin/icu/windea/pls/lang/listeners/ParadoxDefaultGameTypeListeners.kt
@@ -17,6 +17,10 @@ class ParadoxRefreshOnDefaultGameTypeChangedListener : ParadoxDefaultGameTypeLis
             }
         }
 
+        // 刷新分析数据
+        val rootFiles = PlsAnalysisManager.findRootFilesByRootFilePaths(modDirectories)
+        PlsAnalysisManager.refreshAnalysisData(rootFiles)
+
         // 重新解析文件
         val files = PlsAnalysisManager.findAllFilesByRootFilePaths(modDirectories)
         PlsAnalysisManager.reparseFiles(files)